% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_de_correcao_final.R
\name{dirichletBiasCorrection}
\alias{dirichletBiasCorrection}
\title{Correction function in estimation in multivariate Dirichlet regressions.}
\usage{
dirichletBiasCorrection(X, Y)
}
\arguments{
\item{X}{It is the n*k model matrix with the observations in k known covariates}

\item{Y}{It is a random p-vector that has a Dirichlet distribution(λ).}
}
\value{
A list containing:
\item{Estimates}{The values of the beta matrix without correction.}
\item{Standard Errors}{The values of the standard error.}
\item{Bias B(θ)}{The second-order bias vector of theta in matrix format.}
\item{Bias-Corrected Estimates}{Matrix with estimates of the bias correction of θ.}

\description{
Function to calculate the improved estimate in multivariate Dirichlet regressions, based on Melo et al. (2020).
}
