% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcao_de_correcao_final.R
\name{fCoxSnell}
\alias{fCoxSnell}
\title{Correction function in estimation in multivariate Dirichlet regressions.}
\usage{
fCoxSnell(X, Y)
}
\arguments{
\item{X}{It is the n*k model matrix with the observations in k known covariates}

\item{Y}{It is a random p-vector that has a Dirichlet distribution(λ).}
}
\value{
A list containing:
\item{Betas}{The values of the beta matrix without correction.}
\item{Erro Padrão}{The values of the standard error.}
\item{B(θ)}{The second-order bias vector of theta in matrix format.}
\item{θ_BC}{The corrected matrix θ_BC.}
\item{Função Score Corrigida}{The corrected Score function, which is derived from the logarithm of the likelihood function.}
}
\description{
Function to calculate the improved estimate in multivariate Dirichlet regressions, based on Melo et al. (2020).
}
